chore: 优化初始化步骤
